refactor(typescript): reorganize code
