--- packages/typescript/lib/features/completions/directiveComment.ts
@@ -1,72 +1,0 @@
-import type * as vscode from '@volar/language-service';
-import * as nls from 'vscode-nls';
-import type { SharedContext } from '../../types';
-
-const localize = nls.loadMessageBundle(); // TODO: not working
-
-interface Directive {
-	readonly value: string;
-	readonly description: string;
-}
-
-const directives: Directive[] = [
-	{
-		value: '@ts-check',
-		description: localize(
-			'ts-check',
-			"Enables semantic checking in a JavaScript file. Must be at the top of a file.")
-	}, {
-		value: '@ts-nocheck',
-		description: localize(
-			'ts-nocheck',
-			"Disables semantic checking in a JavaScript file. Must be at the top of a file.")
-	}, {
-		value: '@ts-ignore',
-		description: localize(
-			'ts-ignore',
-			"Suppresses @ts-check errors on the next line of a file.")
-	}, {
-		value: '@ts-expect-error',
-		description: localize(
-			'ts-expect-error',
-			"Suppresses @ts-check errors on the next line of a file, expecting at least one to exist.")
-	}
-];
-
-export function register(ctx: SharedContext) {
-	return (uri: string, position: vscode.Position) => {
-
-		const document = ctx.getTextDocument(uri);
-		if (!document)
-			return;
-
-		const prefix = document.getText({
-			start: { line: position.line, character: 0 },
-			end: position,
-		});
-		const match = prefix.match(/^\s*\/\/+\s?(@[a-zA-Z\-]*)?$/);
-		if (match) {
-
-			return directives.map(directive => {
-
-				const item: vscode.CompletionItem = { label: directive.value };
-				item.insertTextFormat = 2 satisfies typeof vscode.InsertTextFormat.Snippet;
-				item.detail = directive.description;
-				const range: vscode.Range = {
-					start: {
-						line: position.line,
-						character: Math.max(0, position.character - (match[1] ? match[1].length : 0)),
-					},
-					end: position,
-				};
-				item.textEdit = {
-					range,
-					newText: directive.value,
-				};
-
-				return item;
-			});
-		}
-		return [];
-	};
-}
