+++ packages/typescript/lib/semanticFeatures/codeActionResolve.ts
@@ -4,24 +4,24 @@ import type { TextDocument } from 'vscode-languageserver-textdocument';
 import { getFormatCodeSettings } from '../configs/getFormatCodeSettings';
 import { getUserPreferences } from '../configs/getUserPreferences';
 import { safeCall } from '../shared';
-import type { SharedContext } from '../types';
+import type { SharedContext } from './types';
 import type { Data, FixAllData, RefactorData } from './codeAction';
-import { fileTextChangesToWorkspaceEdit } from './rename';
+import { convertFileTextChanges } from '../utils/lspConverters';
 
 export function register(ctx: SharedContext) {
 	return async (codeAction: vscode.CodeAction) => {
 
 		const data: Data = codeAction.data;
 		const document = ctx.getTextDocument(data.uri);
-		const [formatOptions, preferences] = document ? await Promise.all([
+		const [formatOptions, preferences] = await Promise.all([
 			getFormatCodeSettings(ctx, document),
 			getUserPreferences(ctx, document),
-		]) : [{}, {}];
+		]);
 
 		if (data?.type === 'fixAll') {
 			resolveFixAllCodeAction(ctx, codeAction, data, formatOptions, preferences);
 		}
-		else if (data?.type === 'refactor' && document) {
+		else if (data?.type === 'refactor') {
 			resolveRefactorCodeAction(ctx, codeAction, data, document, formatOptions, preferences);
 		}
 		else if (data?.type === 'organizeImports') {
@@ -41,7 +41,7 @@ export function resolveFixAllCodeAction(
 ) {
 	const fixes = data.fixIds.map(fixId => safeCall(() => ctx.languageService.getCombinedCodeFix({ type: 'file', fileName: data.fileName }, fixId, formatOptions, preferences)));
 	const changes = fixes.map(fix => fix?.changes ?? []).flat();
-	codeAction.edit = fileTextChangesToWorkspaceEdit(changes, ctx);
+	codeAction.edit = convertFileTextChanges(changes, ctx.fileNameToUri, ctx.getTextDocument);
 }
 
 export function resolveRefactorCodeAction(
@@ -56,7 +56,7 @@ export function resolveRefactorCodeAction(
 	if (!editInfo) {
 		return;
 	}
-	codeAction.edit = fileTextChangesToWorkspaceEdit(editInfo.edits, ctx);
+	codeAction.edit = convertFileTextChanges(editInfo.edits, ctx.fileNameToUri, ctx.getTextDocument);
 	if (editInfo.renameLocation !== undefined && editInfo.renameFilename !== undefined) {
 		codeAction.command = ctx.commands.rename.create(
 			document.uri,
@@ -73,5 +73,5 @@ export function resolveOrganizeImportsCodeAction(
 	preferences: ts.UserPreferences,
 ) {
 	const changes = safeCall(() => ctx.languageService.organizeImports({ type: 'file', fileName: data.fileName }, formatOptions, preferences));
-	codeAction.edit = fileTextChangesToWorkspaceEdit(changes ?? [], ctx);
+	codeAction.edit = convertFileTextChanges(changes ?? [], ctx.fileNameToUri, ctx.getTextDocument);
 }
