+++ packages/typescript/lib/semanticFeatures/types.ts
@@ -3,10 +3,9 @@ import type * as ts from 'typescript';
 import type { TextDocument } from 'vscode-languageserver-textdocument';
 
 export type SharedContext = ServiceContext & {
-	ts: typeof import('typescript');
 	languageServiceHost: ts.LanguageServiceHost;
 	languageService: ts.LanguageService;
-	getTextDocument: (uri: string) => TextDocument | undefined;
+	getTextDocument: (uri: string) => TextDocument;
 	uriToFileName: (uri: string) => string;
 	fileNameToUri: (fileName: string) => string;
 };
